Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int num1, num2; 
+    float num1, num2; 
     char choice;
-    scanf("%d %d %c", &num1, &num2, &choice);
+    scanf("%f %f %c", &num1, &num2, &choice);
     
     switch(choice){
         case '+': printf("%d", num1 + num2); break;
